moved tool route related logic to tool_route_manager
