+++ crates/goose/src/agents/reply_parts.rs
@@ -48,21 +48,16 @@ impl Agent {
         };
 
         // Get tools from extension manager
-        let mut tools = if *self.router_disabled_override.lock().await {
-            // If router is disabled, use regular tools
-            self.list_tools(None).await
-        } else {
-            match tool_selection_strategy {
-                Some(RouterToolSelectionStrategy::Vector) => {
-                    self.list_tools_for_router(Some(RouterToolSelectionStrategy::Vector))
-                        .await
-                }
-                Some(RouterToolSelectionStrategy::Llm) => {
-                    self.list_tools_for_router(Some(RouterToolSelectionStrategy::Llm))
-                        .await
-                }
-                _ => self.list_tools(None).await,
+        let mut tools = match tool_selection_strategy {
+            Some(RouterToolSelectionStrategy::Vector) => {
+                self.list_tools_for_router(Some(RouterToolSelectionStrategy::Vector))
+                    .await
             }
+            Some(RouterToolSelectionStrategy::Llm) => {
+                self.list_tools_for_router(Some(RouterToolSelectionStrategy::Llm))
+                    .await
+            }
+            _ => self.list_tools(None).await,
         };
         // Add frontend tools
         let frontend_tools = self.frontend_tools.lock().await;
